Extract `MockSnapshot`.
